requirements.txt
astroid==2.8.4
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-vision==2.6.1
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.3
isort==5.9.3
labelImg==1.8.6
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
numpy==1.21.3
packaging==21.0
Pillow==8.4.0
platformdirs==2.4.0
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pylint==2.11.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dotenv==0.19.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.2
